# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PulseOps",
    "version": "0.1.0"
  },
  "paths": {
...
    "/test/chat-simulator/status": {
      "get": {
        "tags": ["test-support"],
        "summary": "Chat Simulator Status",
        "operationId": "chat_simulator_status_test_chat_simulator_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSimulatorStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/test/chat-simulator/messages": {
      "get": {
        "tags": ["test-support"],
        "summary": "Chat Simulator Messages",
        "operationId": "chat_simulator_messages_test_chat_simulator_messages_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSimulatorMessagesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/test/chat-simulator/messages/{message_id}/reply": {
      "post": {
        "tags": ["test-support"],
        "summary": "Chat Simulator Reply",
        "operationId": "chat_simulator_reply_test_chat_simulator_messages__message_id__reply_post",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSimulatorReplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSimulatorReplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/test/chat-simulator/state": {
      "delete": {
        "tags": ["test-support"],
        "summary": "Reset Chat Simulator",
        "operationId": "reset_chat_simulator_test_chat_simulator_state_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatSimulatorMessageResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "channel_id": {
            "type": "string",
            "title": "Channel Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "reply_to_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply To Message Id"
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "message_id",
          "channel_id",
          "user_id",
          "direction",
          "text",
          "created_at",
          "metadata"
        ],
        "title": "ChatSimulatorMessageResponse"
      },
      "ChatSimulatorMessagesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatSimulatorMessageResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "ChatSimulatorMessagesResponse"
      },
      "ChatSimulatorReplyRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          }
        },
        "type": "object",
        "required": ["text"],
        "title": "ChatSimulatorReplyRequest"
      },
      "ChatSimulatorReplyResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "processed_message_id": {
            "type": "string",
            "title": "Processed Message Id"
          }
        },
        "type": "object",
        "required": ["message_id", "status", "processed_message_id"],
        "title": "ChatSimulatorReplyResponse"
      },
      "ChatSimulatorStatusResponse": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          }
        },
        "type": "object",
        "required": ["enabled", "tenant_id", "provider", "message_count"],
        "title": "ChatSimulatorStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}